# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AlertPage": {
        "properties": {
          "items": {
            "description": "Записи текущей страницы.",
            "items": {
              "$ref": "#/components/schemas/AlertOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "description": "Размер страницы (1–100, по умолчанию 50).",
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "description": "Количество пропущенных записей, начиная с 0.",
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "description": "Число записей после фильтров, до пагинации.",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "limit",
          "offset",
          "total"
        ],
        "title": "AlertPage",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "description": "Стабильный машинный код для обработки клиентом.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "description": "Подробности ошибки: доступное количество, дата или список ошибок валидации.",
            "title": "Details",
            "type": "object"
          },
          "message": {
            "description": "Человекочитаемое сообщение.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody",
            "description": "Единый объект ошибки API."
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "StockDetail": {
        "properties": {
          "locations": {
            "description": "Объекты с партиями товара. Пустой массив, если партий нет.",
            "items": {
              "$ref": "#/components/schemas/LocationStock"
            },
            "title": "Locations",
            "type": "array"
          },
          "name": {
            "description": "Название товара.",
            "title": "Name",
            "type": "string"
          },
          "sku": {
            "description": "Уникальный артикул товара, например OIL-001.",
            "title": "Sku",
            "type": "string"
          },
          "unit": {
            "description": "Единица измерения количества товара.",
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "sku",
          "name",
          "unit",
          "locations"
        ],
        "title": "StockDetail",
        "type": "object"
      },
...
    }
  },
  "info": {
...
    "title": "Warehouse Purchase Planning API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/alerts": {
      "get": {
        "description": "Коды: stock_shortage, expiring_batch, expired_batch, no_movement, never_moved. Дефицит: available_stock < средний расход × срок поставки, без входящих заказов. Пороги годности и бездействия заданы EXPIRY_WARNING_DAYS и INACTIVITY_DAYS. never_moved имеет location = null и возвращается только без фильтра объекта. Порядок: critical, warning, info; затем SKU, объект, код, партия.",
        "operationId": "listAlerts",
        "parameters": [
          {
            "description": "Фильтр по SKU. Неизвестный SKU — 404.",
            "in": "query",
            "name": "sku",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по SKU. Неизвестный SKU — 404.",
              "examples": [
                "OIL-001"
              ],
              "title": "Sku"
            }
          },
          {
            "description": "Код объекта. Неизвестный код — 404.",
            "in": "query",
            "name": "location",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Код объекта. Неизвестный код — 404.",
              "examples": [
                "MS-01"
              ],
              "title": "Location"
            }
          },
          {
            "description": "Размер страницы; total не зависит от limit.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Размер страницы; total не зависит от limit.",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Количество пропущенных записей после фильтрации.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Количество пропущенных записей после фильтрации.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "items": [
                    {
                      "batch": "MASK-EXPIRED",
                      "code": "expired_batch",
                      "level": "critical",
                      "location": "MS-01",
                      "message": "Партия просрочена: требуется writeoff",
                      "metrics": {
                        "days_to_expiry": -1,
                        "expires_on": "2026-09-22",
                        "remaining_quantity": "4.000",
                        "threshold_days": 30
                      },
                      "sku": "MASK-001"
                    }
                  ],
                  "limit": 50,
                  "offset": 0,
                  "total": 1
                },
                "schema": {
                  "$ref": "#/components/schemas/AlertPage"
                }
              }
            },
            "description": "Страница предупреждений с исходными показателями"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "default": {
                    "summary": "SKU, объект или запрошенная партия не найдены",
                    "value": {
                      "error": {
                        "code": "unknown_sku",
                        "details": {
                          "sku": "UNKNOWN"
                        },
                        "message": "SKU does not exist"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "SKU, объект или запрошенная партия не найдены"
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "default": {
                    "summary": "Невалидные поля или нарушение бизнес-правил",
                    "value": {
                      "error": {
                        "code": "validation_error",
                        "details": {
                          "errors": [
                            {
                              "location": [
                                "body",
                                "quantity"
                              ],
                              "message": "Invalid quantity",
                              "type": "value_error"
                            }
                          ]
                        },
                        "message": "Request validation failed"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Невалидные поля или нарушение бизнес-правил"
          },
          "503": {
            "content": {
              "application/json": {
                "examples": {
                  "default": {
                    "summary": "База временно недоступна",
                    "value": {
                      "error": {
                        "code": "database_unavailable",
                        "details": {},
                        "message": "Database temporarily unavailable"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "База временно недоступна"
          }
        },
        "summary": "Получить предупреждения склада",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/forecast": {
      "post": {
        "description": "Расчёт без изменения БД. Передайте ровно один горизонт: days или months. Месяцы календарные, конец периода включён. При отложенном start_date учитывается дефицит до начала периода. Календарная модель учитывает FEFO, истечение годности и даты неполученных открытых поставок; просроченные поставки исключены. Закупка покрывает непокрытый спрос и страховой запас, округляется до упаковки и минимального заказа. Цена последнего поступления; при её отсутствии стоимость null. stockout_date показана без новой закупки только в пределах горизонта. explanation содержит исходные данные, формулы и ограничения модели.",
        "operationId": "calculateForecast",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "days": {
                  "summary": "На 30 дней со страховым запасом 14 дней",
                  "value": {
                    "days": 30,
                    "location": "MS-01",
                    "safety_stock_days": 14,
                    "sku": "OIL-001"
                  }
                },
                "deferred": {
                  "summary": "Отложенное начало периода",
                  "value": {
                    "location": "MS-01",
                    "months": 3,
                    "safety_stock_days": 14,
                    "sku": "OIL-001",
                    "start_date": "2026-10-01"
                  }
                },
                "months": {
                  "summary": "На три календарных месяца",
                  "value": {
                    "location": "MS-01",
                    "months": 3,
                    "safety_stock_days": 14,
                    "sku": "OIL-001"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/ForecastIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "available_stock": 10,
                  "avg_daily_consumption": 1,
                  "current_stock": 10,
                  "estimated_cost": 439.25,
                  "explanation": {
                    "as_of": "2026-09-23",
                    "assumptions": [
                      "Постоянный спрос; цена последнего поступления; валюта RUB"
                    ],
                    "data_used": [
                      "Расход 90 единиц за 90 полных дней; запас 10"
                    ],
                    "formulas": [
                      "D = 90 / 90 = 1",
                      "ROP = 1 × (7 + 14) = 21",
                      "Qraw = 30 + 14 − 10 = 34",
                      "Q = ceil(max(34, 10) / 5) × 5 = 35"
                    ],
                    "inputs": {
                      "daily_rate_exact": "1",
                      "history_consumption": "90",
                      "lead_time_days": 7,
                      "minimum_order": "10",
                      "pack_size": "5",
                      "required_before_rounding": "34"
                    }
                  },
                  "forecast_demand": 30,
                  "incoming_qty": 0,
                  "location": "MS-01",
                  "name": "Массажное масло",
                  "period": {
                    "days": 30,
                    "from": "2026-09-23",
                    "to": "2026-10-22"
                  },
                  "recommended_order_date": "2026-09-23",
                  "recommended_purchase_qty": 35,
                  "reorder_point": 21,
                  "safety_stock": 14,
                  "sku": "OIL-001",
                  "stockout_date": "2026-10-03",
                  "unit": "л",
                  "unit_price": 12.55,
                  "warnings": [
                    {
                      "code": "below_reorder_point",
                      "level": "warning",
                      "message": "Остаток ниже точки заказа"
                    },
                    {
                      "code": "projected_shortage",
                      "level": "critical",
                      "message": "Без закупки возникнет дефицит"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ForecastOut"
                }
              }
            },
            "description": "Рекомендация и объяснение без размещения заказа"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "default": {
                    "summary": "Некорректный синтаксис JSON",
                    "value": {
                      "error": {
                        "code": "invalid_json",
                        "details": {
                          "errors": [
                            {
                              "location": [
                                "body",
                                1
                              ],
                              "message": "JSON decode error",
                              "type": "json_invalid"
                            }
                          ]
                        },
                        "message": "Malformed JSON"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Некорректный синтаксис JSON"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "default": {
                    "summary": "SKU, объект или запрошенная партия не найдены",
                    "value": {
                      "error": {
                        "code": "unknown_sku",
                        "details": {
                          "sku": "UNKNOWN"
                        },
                        "message": "SKU does not exist"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "SKU, объект или запрошенная партия не найдены"
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "default": {
                    "summary": "Невалидные поля или нарушение бизнес-правил",
                    "value": {
                      "error": {
                        "code": "validation_error",
                        "details": {
                          "errors": [
                            {
                              "location": [
                                "body",
                                "quantity"
                              ],
                              "message": "Invalid quantity",
                              "type": "value_error"
                            }
                          ]
                        },
                        "message": "Request validation failed"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Невалидные поля или нарушение бизнес-правил"
          },
          "503": {
            "content": {
              "application/json": {
                "examples": {
                  "default": {
                    "summary": "База временно недоступна",
                    "value": {
                      "error": {
                        "code": "database_unavailable",
                        "details": {},
                        "message": "Database temporarily unavailable"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "База временно недоступна"
          }
        },
        "summary": "Рассчитать потребность в закупке",
        "tags": [
          "forecast"
        ]
      }
    },
...
    "/api/stock/{sku}": {
      "get": {
        "description": "Детализация по объектам и партиям, включая пустые партии. Партии упорядочены по годности и ID. Известный товар без партий: locations = []. receipt_documents содержит все приходные документы партии.",
        "operationId": "getStockBySku",
        "parameters": [
          {
            "description": "Артикул товара.",
            "in": "path",
            "name": "sku",
            "required": true,
            "schema": {
              "description": "Артикул товара.",
              "examples": [
                "OIL-001"
              ],
              "title": "Sku",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "locations": [
                    {
                      "available_stock": 10,
                      "avg_daily_consumption": 1,
                      "batches": [
                        {
                          "batch": "MANUAL-BATCH-001",
                          "current_stock": 10,
                          "expired": false,
                          "expires_on": "2027-09-23",
                          "receipt_documents": [
                            "MANUAL-RECEIPT-001"
                          ],
                          "received_on": "2026-09-23",
                          "unit_price": 1500
                        }
                      ],
                      "current_stock": 10,
                      "days_of_stock": 10,
                      "expired_stock": 0,
                      "location": "MS-01",
                      "name": "Массажное масло",
                      "nearest_expiry": "2027-09-23",
                      "sku": "OIL-001",
                      "unit": "л"
                    }
                  ],
                  "name": "Массажное масло",
                  "sku": "OIL-001",
                  "unit": "л"
                },
                "schema": {
                  "$ref": "#/components/schemas/StockDetail"
                }
              }
            },
            "description": "Объекты и партии товара"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "default": {
                    "summary": "SKU, объект или запрошенная партия не найдены",
                    "value": {
                      "error": {
                        "code": "unknown_sku",
                        "details": {
                          "sku": "UNKNOWN"
                        },
                        "message": "SKU does not exist"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "SKU, объект или запрошенная партия не найдены"
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "default": {
                    "summary": "Невалидные поля или нарушение бизнес-правил",
                    "value": {
                      "error": {
                        "code": "validation_error",
                        "details": {
                          "errors": [
                            {
                              "location": [
                                "body",
                                "quantity"
                              ],
                              "message": "Invalid quantity",
                              "type": "value_error"
                            }
                          ]
                        },
                        "message": "Request validation failed"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Невалидные поля или нарушение бизнес-правил"
          },
          "503": {
            "content": {
              "application/json": {
                "examples": {
                  "default": {
                    "summary": "База временно недоступна",
                    "value": {
                      "error": {
                        "code": "database_unavailable",
                        "details": {},
                        "message": "Database temporarily unavailable"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "База временно недоступна"
          }
        },
        "summary": "Получить остатки и партии товара",
        "tags": [
          "stock"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Текущий сервер API",
      "url": "/"
    }
  ],
...
}